+++ CHANGELOG.md
@@ -8,6 +8,13 @@ CHANGELOG
 
 - [sdk/dotnet] F# API to specify stack options.
   [#5077](https://github.com/pulumi/pulumi/pull/5077)
+
+- [cli] Disable permalinks to the update details page when using self-managed backends (S3, Azure, GCS). Should the user 
+  want to get permalinks when using a self backend, they can pass a flag:  
+      `pulumi up --suppress-permalink false`.  
+  Permalinks for these self-managed backends will be suppressed on `update`, `preview`, `destroy`, `import` and `refresh` 
+  operations.
+  [#6251](https://github.com/pulumi/pulumi/pull/6251)
   
 ## 2.21.1 (2021-02-18)
 
@@ -56,7 +63,7 @@ CHANGELOG
 
 - [sdk/python] Fixed a bug in `contains_unknowns` where outputs with a property named "values" failed with a TypeError.
   [#6264](https://github.com/pulumi/pulumi/pull/6264)
-
+  
 - [sdk/python] Allowed keyword args in Output.all() to create a dict.
   [#6269](https://github.com/pulumi/pulumi/pull/6269)
   
@@ -65,7 +72,6 @@ CHANGELOG
 
 - [automation/python] Fixed a bug in nested configuration parsing.
   [#6349](https://github.com/pulumi/pulumi/pull/6349)
-
 
 ## 2.20.0 (2021-02-03)